src/traffic_light_classifier/traffic_light_classifier/requirements.txt
numpy==1.21.5
opencv_python==4.10.0.84
pandas==1.3.5
Pillow==9.0.1
rclpy==3.3.14
torch==2.4.1
torchvision==0.19.1
